changed the type of the multiboot arguments in start_kernel() and multiboot()

+++ include/fiwix/kernel.h
@@ -106,8 +106,8 @@ struct kernel_stat {
 extern struct kernel_stat kstat;
 
 unsigned int get_last_boot_addr(unsigned int);
-void multiboot(unsigned long, unsigned long);
-void start_kernel(unsigned long, unsigned long, unsigned int);
+void multiboot(unsigned int, unsigned int);
+void start_kernel(unsigned int, unsigned int, unsigned int);
 void stop_kernel(void);
 void init_init(void);
 void cpu_idle(void);
